+++ meinberlin/apps/bplan/views.py
@@ -1,7 +1,8 @@
+from django.http import HttpResponseRedirect
+from django.urls import reverse
 from django.utils.functional import cached_property
 from django.utils.translation import gettext_lazy as _
 from django.views.generic import DetailView
-from django.views.generic import TemplateView
 
 from adhocracy4.dashboard.blueprints import ProjectBlueprint
 from adhocracy4.dashboard.components.forms.views import ProjectComponentFormView
@@ -10,9 +11,10 @@ from adhocracy4.filters import views as filter_views
 from adhocracy4.filters import widgets as filter_widgets
 from adhocracy4.filters.filters import DefaultsFilterSet
 from adhocracy4.filters.filters import FreeTextFilter
-from meinberlin.apps.bplan import phases as bplan_phases
 from meinberlin.apps.dashboard.mixins import DashboardProjectListGroupMixin
+from meinberlin.apps.extprojects.phases import ExternalPhase
 from meinberlin.apps.extprojects.views import ExternalProjectCreateView
+from meinberlin.apps.projects.utils import get_public_project_url
 
 from . import forms
 from . import models
@@ -32,14 +34,6 @@ class BPlanFilterSet(DefaultsFilterSet):
         fields = ["search"]
 
 
-class BplanStatementSentView(TemplateView):
-    template_name = "meinberlin_bplan/statement_sent.html"
-
-
-class BplanFinishedView(TemplateView):
-    template_name = "meinberlin_bplan/bplan_finished.html"
-
-
 class BplanProjectCreateView(ExternalProjectCreateView):
 
     model = models.Bplan
@@ -55,7 +49,7 @@ class BplanProjectCreateView(ExternalProjectCreateView):
             " to be embedded on external sites."
         ),
         content=[
-            bplan_phases.StatementPhase(),
+            ExternalPhase(),
         ],
         image="",
         settings_model=None,
@@ -98,22 +92,14 @@ class BplanProjectDispatchMixin(DetailView):
     def project(self):
         return self.get_object()
 
-    @cached_property
-    def module(self):
-        if self.project.published_modules.count() == 1:
-            return self.project.published_modules.first()
-        elif len(self.project.running_modules) == 1:
-            return self.project.running_modules[0]
-
     def dispatch(self, request, *args, **kwargs):
+        # Bplans live on an external platform (Diplan) and the on-site detail
+        # page offers no participation, so redirect to the external project URL.
+        # Drafts may not have a url yet; send those to the dashboard edit page.
         if self.project.project_type == "meinberlin_bplan.Bplan":
-            kwargs["project"] = self.project
-            kwargs["module"] = self.module
-
-            return self._dispatch_bplan_module_view()(request, *args, **kwargs)
+            url = get_public_project_url(self.project) or reverse(
+                "a4dashboard:project-edit",
+                kwargs={"project_slug": self.project.slug},
+            )
+            return HttpResponseRedirect(url)
         return super().dispatch(request, *args, **kwargs)
-
-    def _dispatch_bplan_module_view(self):
-        if self.module and self.module.last_active_phase:
-            return self.module.last_active_phase.view.as_view()
-        return super().dispatch