+++ meinberlin/apps/projects/serializers.py
@@ -12,6 +12,7 @@ from adhocracy4.projects.models import Project
 from adhocracy4.projects.models import Topic
 from meinberlin.apps import logger
 from meinberlin.apps.kiezradar.models import ProjectType
+from meinberlin.apps.projects.utils import get_public_project_url
 
 
 class CommonFields:
@@ -66,7 +67,6 @@ class ProjectSerializer(
     type = serializers.SerializerMethodField()
     url = serializers.SerializerMethodField()
     topics = serializers.SerializerMethodField()
-    identifier = serializers.SerializerMethodField()
 
     def __init__(self, *args, **kwargs):
         self.now = kwargs.pop("now", timezone.now())
@@ -91,7 +91,6 @@ class ProjectSerializer(
             "description",
             "district",
             "future_phase",
-            "identifier",
             "organisation",
             "participation",
             "participation_active",
@@ -155,12 +154,7 @@ class ProjectSerializer(
         return instance.name
 
     def get_url(self, instance):
-        if instance.project_type in (
-            "meinberlin_extprojects.ExternalProject",
-            "meinberlin_bplan.Bplan",
-        ):
-            return instance.externalproject.url
-        return instance.get_absolute_url()
+        return get_public_project_url(instance)
 
     def get_tile_image(self, instance):
         image_url = ""
@@ -259,11 +253,6 @@ class ProjectSerializer(
 
     def get_cost(self, instance):
         return ""
-
-    def get_identifier(self, instance):
-        if hasattr(instance, "externalproject"):
-            if hasattr(instance.externalproject, "bplan"):
-                return instance.externalproject.bplan.identifier
 
 
 class ActiveProjectSerializer(ProjectSerializer):